Use TCHAR in native bootstrapper
We wish to share the source for the native bootstrapper between Windows and *NIX.  The first step is doing so is to move away from WCHAR, since it is not portable across platforms.  Instead, we'll use TCHAR, which we'll define to just be char on *NIX platforms.  This forces developers in klr to not assume that a character is two bytes.

At runtime, there's no change at runtime. With _UNICODE defined, TCHAR is alias for WCHAR, LPCTSTR is an alias for LPCWSTR, etc.